+++ packages/feflow-cli/src/core/native/help.ts
@@ -17,7 +17,7 @@ const showHelp = (commands: Array<Object>) => {
   const sections = [
     {
       header: 'Usage',
-      content: '$ feflow [options] [command]'
+      content: '$ fef [options] [command]'
     },
     {
       header: 'Commands',
@@ -38,6 +38,14 @@ const showHelp = (commands: Array<Object>) => {
         {
           name: 'help',
           description: 'Print this help and exit successfully.'
+        },
+        {
+          name: 'disable-check',
+          description: 'Disable @feflow/cli and installed plugins check update'
+        },
+        {
+          name: 'auto-update',
+          description: 'Auto update @feflow/cli and installed plugins'
         }
       ]
     }